ADD shared file between cpp and glsl
