+++ shaders/rasterization/hud/item_icon_shader.frag
@@ -1,11 +1,13 @@
 #version 450
 
+#include "common.glsl"
+
 layout (set = 0, binding = 0) uniform sampler2DArray image;
 
 layout(push_constant) uniform PushConstants
 {
-	int index;
-} pc;
+	ItemIconPushConstant pc;
+};
 
 layout (location = 0) in vec2 inUV;
 
@@ -13,5 +15,5 @@ layout (location = 0) out vec4 outColor;
 
 void main()
 {
-	outColor = texture(image, vec3(inUV, pc.index));
+	outColor = texture(image, vec3(inUV, pc.layer));
 }
