+++ shaders/rasterization/sky/sun.vert
@@ -1,17 +1,16 @@
 #version 450
 
-// Vertex shader for the skybox
+#include "common.glsl"
 
 layout(set = 0, binding = 0) uniform CameraMatrices
 {
-	mat4 view;
-	mat4 projection;
-}cm;
+	ViewProjMatrices cm;
+};
 
 layout(push_constant) uniform PushConstants
 {
-	mat4 model;
-}pc;
+	ModelMatrice pc;
+};
 
 layout(location = 0) in vec3 position;
 layout(location = 1) in vec3 texCoord;
@@ -21,7 +20,7 @@ layout(location = 0) out vec3 rayDir;
 
 void main()
 {
-	vec4 pos = cm.projection * cm.view * pc.model * vec4(position, 1.0);
+	vec4 pos = cm.proj * cm.view * pc.model * vec4(position, 1.0);
 	gl_Position = pos.xyww;
 
 	rayDir = normalize(position);
