+++ shaders/rasterization/water_shader.vert
@@ -1,15 +1,16 @@
 #version 450
 
+#include "common.glsl"
+
 layout(set = 0, binding = 0) uniform CameraMatrices
 {
-	mat4 view;
-	mat4 projection;
-}cm;
+	ViewProjMatrices cm;
+};
 
 layout(push_constant) uniform PushConstants
 {
-	mat4 model;
-}pc;
+	ModelMatrice pc;
+};
 
 layout(location = 0) in vec3 positions;
 layout(location = 1) in vec3 normal;
@@ -24,7 +25,7 @@ layout(location = 2) out vec4 fragPosWorldSpace;
 
 void main()
 {
-	gl_Position = cm.projection * cm.view * pc.model * vec4(positions, 1.0);
+	gl_Position = cm.proj * cm.view * pc.model * vec4(positions, 1.0);
 
 	fragNormal = normal;
 	fragTexCoords = vec3(texCoords, texLayer);
